app/requirements.txt
cachetools==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.0.3 ; python_version >= "3.11" and python_version < "4.0"
cloudevents==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
flask==2.0.2 ; python_version >= "3.11" and python_version < "4"
functions-framework==3.0.0 ; python_version >= "3.11" and python_version < "4"
google-api-core==2.11.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.11.1 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.17.3 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-appengine-logging==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-audit-log==0.2.5 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-datastore==2.15.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-logging==3.6.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.60.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.60.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.57.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.57.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.11" and python_version < "4.0"
pathtools==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.24.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.0.6 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
setuptools==60.3.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
watchdog==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
